+++ src/api/spl.py
@@ -325,17 +325,15 @@ def verify_token(token_params):
     return False
 
 
-def get_owned_resource_sum(account, resource, token_params):
+def get_owned_resource_sum(account, resource):
     address = land_url + 'land/resources/owned'
-
-    params = token_params
-    params['player'] = account
-    params['resource'] = resource
+    params = {'player': account, 'resource': resource}
 
     result = http.get(address, params=params).json()
     if result and 'data' in result:
         df = pd.DataFrame(result['data'])
-        return df.amount.sum()
+        if 'amount' in df.columns.tolist():
+            return df.amount.sum()
     return None
 
 
@@ -346,6 +344,15 @@ def get_prices():
 
 def spl_get_pools():
     address = land_url + 'land/liquidity/pools'
+
+    result = http.get(address).json()
+    if result and 'data' in result:
+        return pd.DataFrame(result['data'])
+    return pd.DataFrame()
+
+
+def get_liquidity(account, resource):
+    address = land_url + 'land/liquidity/pools/' + str(account) + '/' + resource
 
     result = http.get(address).json()
     if result and 'data' in result: